+++ aqua/diagnostics/ecmean/cli_ecmean.py
@@ -51,6 +51,7 @@ def parse_arguments(arguments):
 def reader_data(model, exp, source,
                 catalog=None, regrid='r100',
                 keep_vars=None, loglevel='WARNING',
+                startdate=None, enddate=None,
                 reader_kwargs: dict = {}):
     """
     Simple function to retrieve and do some operation on reader data
@@ -61,6 +62,8 @@ def reader_data(model, exp, source,
         source (str): source of the data
         catalog (str, optional): catalog to be used, defaults to None
         regrid (str, optional): regrid method, defaults to 'r100'
+        startdate (str, optional): start date in the format YYYY-MM-DD, defaults to None
+        enddate (str, optional): end date in the format YYYY-MM-DD, defaults to None
         keep_vars (list, optional): list of variables to keep, defaults to None
         loglevel (str, optional): logging level, defaults to 'WARNING'
         reader_kwargs (dict, optional): list of reader_kwargs. Defaults to {}.
@@ -78,21 +81,25 @@ def reader_data(model, exp, source,
     # Try to read the data, if dataset is not available return None
     try:
         reader = Reader(
-            model=model, exp=exp, source=source, catalog=catalog, 
+            model=model, exp=exp, source=source, catalog=catalog,
             regrid=regrid, **reader_kwargs
         )
-        xfield = reader.retrieve()
-        if regrid is not None:
-            xfield = reader.regrid(xfield)
-
+        xfield = reader.retrieve(startdate=startdate, enddate=enddate, var=keep_vars)
     except Exception as err:
         reader_logger.error('Error while reading model %s: %s', model, err)
         return None
+    
+    # regrid after variable selection
+    if regrid is not None:
+        try: 
+            return reader.regrid(xfield)
+        except Exception as err:
+            reader_logger.error('Error while regridding model %s: %s', model, err)
+            return None
 
-    # return only vars that are available: slower but avoid reader failures
-    if keep_vars is None:
-        return xfield
-    return xfield[[value for value in keep_vars if value in xfield.data_vars]]
+
+
+
 
 def data_check(data_atm, data_oce, logger=None):
     """
@@ -168,8 +175,10 @@ def set_title(diagnostic: str, model: str, exp: str,
     """
     if diagnostic == 'performance_indices':
         diag_name = 'Performance Indices'
-    if diagnostic == 'global_mean':
+    elif diagnostic == 'global_mean':
         diag_name = 'Global Mean Bias'
+    else:
+        raise ValueError(f"Unknown diagnostic {diagnostic} for title generation")
         
     builder = TitleBuilder(
         diagnostic=diag_name,
@@ -253,8 +262,7 @@ if __name__ == '__main__':
     # define the output properties
     outputdir = output_config.get('outputdir')
     rebuild = output_config.get('rebuild', True)
-    save_pdf = output_config.get('save_pdf', False)
-    save_png = output_config.get('save_png', False)
+    save_format = output_config.get('save_format', [])
 
     # merge config args works only with a predefined set of options, need to extend it
     numproc = get_arg(args, 'nprocs', ecmean_config.get('nprocs', 1))
@@ -316,11 +324,13 @@ if __name__ == '__main__':
             logger.info('Loading atmospheric data %s', model)
             data_atm = reader_data(model=model, exp=exp, source=source_atm,
                                    catalog=catalog, keep_vars=atm_vars, regrid=regrid,
+                                   startdate=startdate, enddate=enddate,
                                    reader_kwargs=reader_kwargs)
 
             logger.info('Loading oceanic data from %s', model)
             data_oce = reader_data(model=model, exp=exp, source=source_oce,
                                    catalog=catalog, keep_vars=oce_vars, regrid=regrid,
+                                   startdate=startdate, enddate=enddate,
                                    reader_kwargs=reader_kwargs)
 
             # check the data
@@ -359,14 +369,9 @@ if __name__ == '__main__':
                 ecmean.store(yamlfile=filename_dict['yml'], tablefile=filename_dict['txt'])
             ecmean_fig = ecmean.plot(diagname=diagnostic, returnfig=True, storefig=False)
 
-            if save_pdf:
-                logger.info('Saving PDF %s plot...', diagnostic)
-                outputsaver.save_pdf(fig=ecmean_fig, diagnostic_product=diagnostic,
-                                     metadata=metadata, rebuild=rebuild)
-
-            if save_png:
-                logger.info('Saving PNG %s plot...', diagnostic)
-                outputsaver.save_png(fig=ecmean_fig, diagnostic_product=diagnostic,
-                                     metadata=metadata, rebuild=rebuild)
+            if save_format:
+                logger.info("Saving ecmean %s plot in format(s): %s", diagnostic, save_format)
+                outputsaver.save_figure(fig=ecmean_fig, diagnostic_product=diagnostic,
+                                        metadata=metadata, rebuild=rebuild, extension=save_format)
 
             logger.info('ECmean4 diagnostic completed.')