+++ aqua/diagnostics/ensemble/plot_ensemble_zonal.py
@@ -3,7 +3,7 @@ import xarray as xr
 from aqua.core.exceptions import NoDataError
 from aqua.core.graphics import plot_vertical_profile
 from aqua.core.logger import log_configure
-from aqua.diagnostics.base import TitleBuilder
+from aqua.diagnostics.base import TitleBuilder, SAVE_FORMAT
 from aqua.core.util import find_vert_coord
 
 from .base import BaseMixin
@@ -93,8 +93,7 @@ class PlotEnsembleZonal(BaseMixin):
         title_std=None,
         figure_size=[10, 8],
         cbar_label=None,
-        save_pdf=True,
-        save_png=True,
+        save_format=SAVE_FORMAT,
         dpi=300,
         units=None,
         ylim=(5500, 0),
@@ -120,8 +119,7 @@ class PlotEnsembleZonal(BaseMixin):
             title_std (str, optional): Title for the standard deviation plot. Auto-generated if None.
             figure_size (list[int], optional): Figure size [width, height]. Default is [10, 8].
             cbar_label (str, optional): Label for the colorbar.
-            save_pdf (bool, optional): Save plots as PDF. Default is True.
-            save_png (bool, optional): Save plots as PNG. Default is True.
+            save_format (str or list, optional): Format(s) to save plots in (e.g. 'png', 'pdf', 'svg'). Default is SAVE_FORMAT.
             dpi (int, optional): Resolution for saved figures. Default is 300.
             units (str, optional): Units of the variable. Used in titles and labels if provided.
             ylim (tuple, optional): Y-axis limits for the plot (vertical levels). Default is (5500, 0).
@@ -225,9 +223,7 @@ class PlotEnsembleZonal(BaseMixin):
         self.logger.debug(f"Saving Lev-Lon Zonal-average ensemble-STD as pdf and png")
 
         # Saving plots
-        if save_png:
-            self.save_figure(var=var, fig=fig1, fig_std=fig2, description=description, format="png", dpi=dpi)
-        if save_pdf:
-            self.save_figure(var=var, fig=fig1, fig_std=fig2, description=description, format="pdf")
+        self.save_figure(
+            var=var,fig=fig1, fig_std=fig2, description=description, format=save_format, dpi=dpi)
 
         return {"mean_plot": [fig1, ax1], "std_plot": [fig2, ax2]}